add workflow

+++ src/Controller/CategoryController.php
@@ -76,6 +76,12 @@ class CategoryController extends AbstractController
     #[Route('/{id}', name: 'app_category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if(count($category->getAdverts())>0)  {
+            $error = "Impossible de supprimer cette categorie car des annonces y sont rattachées";
+            $this->addFlash('error',$error);
+            return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
             $categoryRepository->remove($category, true);
         }
